First iteration to allow mulitple connections
This modifies the original behavior of the library and how it is
consumed to allow connections to mulitple databases and warehouses.
Instead of connecting to a database automatically, the consuming
application must now add a Snowflex.ConnectionPool to its
supervision tree. Many connection pools can be started.

The connection pool must also be specified when performing a query, so
that Snowflex knows which connection pool to use.