fix type errors
The child_spec was initializing the worker with a tuple instead of a
propslist.

Also, the Snowflex module was referencing two types:

  * `:odbc.odbc_data_type`
  * `:odbc.value`

These types are defined in the odbc OTP documentation, but do not
actually exist in the OTP source code.

This change corrects the child_spec call and steals the rest of the
pertinent odbc type documentation and implements it for ourselves.

Dialyzer has been installed to verify these changes, however the
child_spec change, which is part of the Connection __using__ callback,
must be (and has been) verified in a consuming library.