Handle utf-16 encoding gracefully when querying Snowflake
This stemmed from an issue we saw in production, a result we were
getting included some garbled characters because of a UTF-8 and UTF-16
difference.

We were getting the string:

```
"CVS PharmacyÂ®"
```

Now, Â = U+00C2 and ® = U+00AE in UTF-8.

However, what we really wanted was:

```
"CVS Pharmacy®"
```

As the Â character is a mistaken extra byte.

Here, we introduce a function to properly handle this and return the
desired results using the latin1 encoding with UTF-8.

I have tried a couple ways of going about testing this, but the coverage
so far does not allow me an easy way to even hit this branch of the
functions in result or snowflex itself. I will confer with someone else
to get a real test written.