feat(test): sqlite schema testing
We add a migration generator module which provides a macro to generate the migrations needed to test any Ecto Schema.
This allows you to swap in another Ecto adapter for local testing without connecting to Snowflake.

To use this you can bring the Snowflex.SQLiteTestRepo into your project and use it as you normally would for testing.
You will need to drop and create the repo before starting the tests each time so the migrations can work from a blank slate.

You will also need to run the generate_migrations macro in your test_helper files with a list of all the Ecto Schemas you wish to test.