chore: add timeout to disconnect
DAPPS-4165

## Context

Timeout for disconnect currently defaults to 60 seconds with no way to configure it. This is because the function `opts` are not passed to the `disconnect` handler.

## Decision

Add `func_opts` to the `state` to allow us access to the passed in `opts` in the `disconnect`  handler.

## Consequences

Disconnect timeout will now respect the passed in timeout from either the `Repo.transaction/2` call, the runtime configuration, or Ecto’s default timeout.

## Related to

https://github.com/pepsico-ecommerce/dpi/pull/3923